Add ways to get all the days in the period
Have changed it from using the session date as the schedule to
using the Monday of the period end. Reason being the actual period
is from Monday to Monday. If we do it from Saturday then we count
the day after the session (Sunday) as part of the next session period,
which is not what we want. Now we're correctly counting it as part
of the current cycle.